Bullet: fixed compile error on buggy Mingw

+++ src/3rdparty/physics/bullet/src/LinearMath/btScalar.h
@@ -440,6 +440,7 @@ inline int btGetVersion()
 	SIMD_FORCE_INLINE btScalar btFmod(btScalar x, btScalar y) { return fmod(x, y); }
 
 #else//BT_USE_DOUBLE_PRECISION
+#include <cmath>
 
 	SIMD_FORCE_INLINE btScalar btSqrt(btScalar y)
 	{
@@ -469,20 +470,20 @@ inline int btGetVersion()
 		return x * y;
 	#endif
 	#else
-		return sqrtf(y);
+		return std::sqrt(y);
 	#endif
 	}
-	SIMD_FORCE_INLINE btScalar btFabs(btScalar x) { return fabsf(x); }
-	SIMD_FORCE_INLINE btScalar btCos(btScalar x) { return cosf(x); }
-	SIMD_FORCE_INLINE btScalar btSin(btScalar x) { return sinf(x); }
-	SIMD_FORCE_INLINE btScalar btTan(btScalar x) { return tanf(x); }
+	SIMD_FORCE_INLINE btScalar btFabs(btScalar x) { return std::fabs(x); }
+	SIMD_FORCE_INLINE btScalar btCos(btScalar x) { return std::cos(x); }
+	SIMD_FORCE_INLINE btScalar btSin(btScalar x) { return std::sin(x); }
+	SIMD_FORCE_INLINE btScalar btTan(btScalar x) { return std::tan(x); }
 	SIMD_FORCE_INLINE btScalar btAcos(btScalar x)
 	{
 		if (x < btScalar(-1))
 			x = btScalar(-1);
 		if (x > btScalar(1))
 			x = btScalar(1);
-		return acosf(x);
+		return std::acos(x);
 	}
 	SIMD_FORCE_INLINE btScalar btAsin(btScalar x)
 	{
@@ -490,14 +491,14 @@ inline int btGetVersion()
 			x = btScalar(-1);
 		if (x > btScalar(1))
 			x = btScalar(1);
-		return asinf(x);
+		return std::asin(x);
 	}
-	SIMD_FORCE_INLINE btScalar btAtan(btScalar x) { return atanf(x); }
-	SIMD_FORCE_INLINE btScalar btAtan2(btScalar x, btScalar y) { return atan2f(x, y); }
-	SIMD_FORCE_INLINE btScalar btExp(btScalar x) { return expf(x); }
-	SIMD_FORCE_INLINE btScalar btLog(btScalar x) { return logf(x); }
-	SIMD_FORCE_INLINE btScalar btPow(btScalar x, btScalar y) { return powf(x, y); }
-	SIMD_FORCE_INLINE btScalar btFmod(btScalar x, btScalar y) { return fmodf(x, y); }
+	SIMD_FORCE_INLINE btScalar btAtan(btScalar x) { return std::atan(x); }
+	SIMD_FORCE_INLINE btScalar btAtan2(btScalar x, btScalar y) { return std::atan2(x, y); }
+	SIMD_FORCE_INLINE btScalar btExp(btScalar x) { return std::exp(x); }
+	SIMD_FORCE_INLINE btScalar btLog(btScalar x) { return std::log(x); }
+	SIMD_FORCE_INLINE btScalar btPow(btScalar x, btScalar y) { return std::pow(x, y); }
+	SIMD_FORCE_INLINE btScalar btFmod(btScalar x, btScalar y) { return std::fmod(x, y); }
 
 #endif//BT_USE_DOUBLE_PRECISION
